Create httpvalidate.py
U will of course have to make a few changes as needed but that should do it. Make a GitHub project name it whatever, then create a new file. Name it whatever just make sure it's .py extension. Such as whatever.py then paste the script in the file. Save it. Clone your repository to your machine. Extract your script. Open python on your machine. Install the three requirements at the very top of the script with pip. Then run the script. Choose a number for your test of choice. Input your url etc and you in the test. 

Some tests will return errors for a few reasons but I'll let you determine what and why as this is a great baseline. 

In the future you'll want to use lists and minimize the script wherever possible as this is quite a large script with not the best commentary on what is what.